+++ Chapter16/Chapter16-1-2/Program16-1-2.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Threading.Tasks;
@@ -22,42 +21,18 @@ namespace Chapter16_1_2 {
 
             // 並列処理をする場合
             var wStopWatch = Stopwatch.StartNew();
-            var wSearchFileTasks = new List<Task>();
-            foreach (string wCsFile in wAllCsFiles) {
-                wSearchFileTasks.Add(Task.Run(() => SearchCsFiles(wCsFile)));
-            }
-            await Task.WhenAll(wSearchFileTasks);
+            var wParallelProcessor = new ParallelFileProcessor();
+            await wParallelProcessor.ProcessFilesAsync(wAllCsFiles);
             wStopWatch.Stop();
             Console.WriteLine($"実行時間（並列処理をする場合）：{wStopWatch.ElapsedMilliseconds}ミリ秒");
 
             // 並列処理をしない場合
             wStopWatch = Stopwatch.StartNew();
-            foreach (string wCsFile in wAllCsFiles) {
-                SearchCsFiles(wCsFile);
-            }
+            var wSequentialProcessor = new SequentialFileProcessor();
+            wSequentialProcessor.ProcessFiles(wAllCsFiles);
             wStopWatch.Stop();
             Console.WriteLine($"実行時間（並列処理をしない場合）：{wStopWatch.ElapsedMilliseconds}ミリ秒");
 
-        }
-        /// <summary>
-        /// 「async」と「await」の両方を含むファイルを検索するメソッド
-        /// </summary>
-        /// <param name="vCsFile">C#のソースファイル</param>
-        static void SearchCsFiles(string vCsFile) {
-            try {
-                var wCsFileContents = File.ReadAllText(vCsFile);
-                if (wCsFileContents.Contains("async") && wCsFileContents.Contains("await")) {
-                    Console.WriteLine(Path.GetFullPath(vCsFile));
-                }
-            } catch (UnauthorizedAccessException wEx) {
-                Console.WriteLine($"ファイルへのアクセスが拒否されました：{wEx.Message}");
-            } catch (PathTooLongException wEx) {
-                Console.WriteLine($"ファイルパスが長すぎます：{wEx.Message}");
-            } catch (OutOfMemoryException wEx) {
-                Console.WriteLine($"メモリが不足しています：{wEx.Message}");
-            } catch (Exception wEx) {
-                Console.WriteLine($"エラーが発生しました：{wEx.Message}");
-            }
         }
     }
 }
